+++ src/lib/pages/interact/component/drawer/body/ModuleSelectMainBody.tsx
@@ -141,7 +141,7 @@ export const ModuleSelectMainBody = ({
               closeModal();
             }
           }}
-          disabled={!selectedModule}
+          isDisabled={!selectedModule}
         >
           Select this module
         </Button>